Move header and footer to layout folder
