+++ components/layout/index.tsx
@@ -1,5 +1,6 @@
-import { Footer, Header } from 'components';
 import { Fragment, ReactNode } from 'react';
+import { Footer } from './footer';
+import { Header } from './header';
 
 type Props = {
   children: ReactNode;
